+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - The citation file.
+- Link to the CWA dataset.
 
 ### Changed
 
@@ -20,9 +21,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Consistent handling of single GPU runs in DistributedManager
+
 ### Security
 
 ### Dependencies
+
+- Updated DGL build in Dockerfile
+- Updated default base image
+- Moved Onnx from optional to required dependencies
+- Optional Makani dependency required for SFNO model.
 
 ## [0.5.0] - 2024-01-25